feat: add economy and ship utility modules
