+++ src/game/shop.ts
@@ -1,6 +1,52 @@
-import { type Part } from '../config/parts'
+import { type Part, ALL_PARTS, RARE_PARTS } from '../config/parts'
 import { nextTierCost, ECONOMY } from '../config/economy'
-import { rollInventory, getEconomyModifiers } from './index'
+import { getEconomyModifiers } from './economy'
+
+let RARE_TECH_CHANCE = 0.1;
+export function setRareTechChance(ch:number){ RARE_TECH_CHANCE = ch; }
+export function getRareTechChance(){ return RARE_TECH_CHANCE; }
+
+export function tierCap(research:{Military:number, Grid:number, Nano:number}): Record<'Military'|'Grid'|'Nano', number>{
+  return {
+    Military: Math.max(1, Math.min(3, Math.floor(research.Military || 1))),
+    Grid: Math.max(1, Math.min(3, Math.floor(research.Grid || 1))),
+    Nano: Math.max(1, Math.min(3, Math.floor(research.Nano || 1))),
+  };
+}
+
+export function rollInventory(research:{Military:number, Grid:number, Nano:number}, count: number = ECONOMY.shop.itemsBase){
+  const capByCat = tierCap(research);
+
+  // Filter parts pool to only include parts within research tier limits and exclude rares
+  const pool = ALL_PARTS.filter((p:Part) =>
+    !p.rare && p.tier === capByCat[p.tech_category as 'Military'|'Grid'|'Nano']
+  );
+
+  const items:Part[] = [];
+
+  const available = [...pool];
+  const rareAvailable = [...RARE_PARTS];
+  while(items.length < count){
+    const useRare = Math.random() < RARE_TECH_CHANCE;
+    if(useRare && rareAvailable.length > 0){
+      const ridx = Math.floor(Math.random() * rareAvailable.length);
+      items.push(rareAvailable[ridx]);
+      rareAvailable.splice(ridx,1);
+      continue;
+    }
+    if(available.length > 0){
+      const idx = Math.floor(Math.random() * available.length);
+      const part = available[idx];
+      items.push(part);
+      available.splice(idx,1);
+    } else if(pool.length > 0){
+      const idx = Math.floor(Math.random() * pool.length);
+      const part = pool[idx];
+      items.push(part);
+    } else break;
+  }
+  return items;
+}
 
 export function doRerollAction(resources:{credits:number}, rerollCost:number, research:{Military:number, Grid:number, Nano:number}){
   if(resources.credits < rerollCost) return { ok:false as const };
